Moves 2 fulp TGUIs to fulp TGUI folder
